Make Karma conversation only
Remove library functions and reduce karma to the conversation needed for
the study.
Update the README to conversation only
