+++ README.md
@@ -1,16 +1,10 @@
-# Karma
-Bot orchestration framework
-
 ## Description
-Experience Sampling Assessment bot.
+Experience Sampling Assessment bot based on <lib-name> and [Botkit].
 
 ### Target platforms
 Current:
  - Facebook messenger
 
-## Documentation
- * [Installation](docs/Install.md)
- * [Testing](docs/Testing.md)
-
 # License
+[Botkit](https://github.com/howdyai/botkit/)
 [BSD-2-Clause](LICENSE)
