--- docs/Testing.md
@@ -1,14 +1,0 @@
-# Testing
-
-## Tests
-no tests yet
-
-## Linting
-
-```
-# install eslint
-npm install eslint --save-dev
-
-# run eslint and let it fix your errors remove the "--fix" to fix errors yourself
-yarn lint --fix
-```
